requirements.txt
flask
flask-sqlalchemy
dotenv
flask-wtf
flask-login
wtforms
bleach
bcrypt
flask-bcrypt
cryptography
email-validator
flask-security-too